requirements.txt
asgiref
astroid
autopep8
boto3
botocore
certifi
chardet
colorama
defusedxml
Django
django-allauth
django-countries
django-crispy-forms
django-debug-toolbar
django-filter
django-storages
djangorestframework
idna
isort
jmespath
lazy-object-proxy
Markdown
mccabe
oauthlib
pep8
Pillow
psycopg2
pycodestyle
pylint
python-dateutil
python-decouple
python3-openid
pytz
requests
requests-oauthlib
s3transfer
six
sqlparse
stripe
toml
urllib3
wrapt
